Alfresco Summit 2014 - Alfresco - So Much Better with a Lisp!
Recorded demo for the lightning talk "Alfresco - So Much Better with a Lisp!". This is the version that was presented during the lighting talk and does not include any comments or annotations describing what's going on. It may be better to view the annotated version ( https://asciinema.org/a/12306/ ) instead.