Router TUI
This is a TUI based on nodejs and blessed (a curses-like library). I ship out a fair number of routers. Verifying data, configuring the routers, and processing all the paperwork involved was very tedious, so I created a portal to streamline the process.

The web-based GUI provides a place to add sites and review deployed router data. This TUI runs on my LAN and automates the process of preparing a router for shipment. It loads a firmware to the router and configures the router for the site. It also uses the UPS API to prepare a label for shipment based on the desired speed of service; it then creates a PDF of the shipping label and spools it to my printer.